+++ Source/CesiumEditor/Private/IonQuickAddPanel.cpp
@@ -8,6 +8,7 @@
 #include "CesiumIonRasterOverlay.h"
 #include "Editor.h"
 #include "PropertyCustomizationHelpers.h"
+#include "SelectCesiumIonToken.h"
 #include "Styling/SlateStyle.h"
 #include "Widgets/Images/SThrobber.h"
 #include "Widgets/Input/SButton.h"
@@ -177,7 +178,21 @@ void IonQuickAddPanel::AddIonTilesetToLevel(TSharedRef<QuickAddItem> item) {
         TEXT("Cannot add an ion asset without an active connection"));
     return;
   }
-  connection->asset(item->tilesetID)
+
+  std::vector<int64_t> assetIDs{item->tilesetID};
+  if (item->overlayID > 0) {
+    assetIDs.push_back(item->overlayID);
+  }
+
+  SelectCesiumIonToken::SelectAndAuthorizeToken(assetIDs)
+      .thenInMainThread([connection, tilesetID = item->tilesetID](
+                            const std::optional<Token>& /*maybeToken*/) {
+        // If token selection was canceled, or if an error occurred while
+        // selecting the token, ignore it and create the tileset anyway. It's
+        // already been logged if necessary, and we can let the user sort out
+        // the problem using the resulting Troubleshooting panel.
+        return connection->asset(tilesetID);
+      })
       .thenInMainThread([item, connection](Response<Asset>&& response) {
         if (!response.value.has_value()) {
           return connection->getAsyncSystem().createResolvedFuture<int64_t>(