+++ Source/CesiumRuntime/CesiumRuntime.Build.cs
@@ -139,7 +139,8 @@ public class CesiumRuntime : ModuleRules
                 "LevelSequence",
                 "Projects",
                 "RenderCore",
-                "SunPosition"
+                "SunPosition",
+                "DeveloperSettings"
             }
         );